// src/interface/fournisseur.interface.ts
export interface FournisseurInterface {
  name: string;
  adress: string;
  zipCode: string;
  ville: string;
  telephone: string;
  //productPrices?: ProductPrice[];
}
export interface ProductPrice {
  id: number;
  prix: number;
  productId: number;
  fournisseurId: number;
  //product?: Product;
  fournisseur?: FournisseurInterface;
}